server: split off password handling from ldap_handler
